+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 - *Mapsforge General Settings* to *Settings / Maps*. To control `lineScale`, `textScale` and `symbolScale` of all mapsforge layers. Changes require a restart of the app.
 - Add some debug/commit information to *About* page in app debug mode.
+- Update routine on app start. Compares current version with last installed version and maybe run several updates until the settings structure meets the structure of the current version.
 - More source options for layer `online-raster-xyz`:
     - *Esri World StreetMap*
     - *Esri World TopoMap*